Fix power manager service shutdown issue
:Release Notes:
Fix power manager segmentation fault issue

:Detailed Notes:
- Fix delay in the shutdown of service when command
'systemctl stop powermgr' executed
- Fix segmentation fault during the service stop
The root cause of issue is LSGmainDetach() used along with
LSUnregister() function. Due that LsHandle corrupted,
which in-turn causing the delay in stopping the service
and delay in the reboot
Trying to delete pointer two times in the destructor of
class which causes segmentation fault for mpconfig pointer

:Testing Performed:
Tested on reference build 1819.
test case :QR-228

:QA Notes:
N/A

:Issues Addressed:
[PLAT-84035] [Webos PRO] : Delayed Boot with stop job logs observed.

Change-Id: I81d03db8b99200a6f6c533d77daf8cb35ba60a45
Reviewed-on: http://gpro.lge.com/256073
Reviewed-by: AnilKumar Reddy <anilkumar.reddy@lge.com>
Reviewed-by: Sushovan Gangopadhy <sushovan.g@lge.com>
Tested-by: Punnarao D <punnarao.d@lge.com>
Reviewed-by: Punnarao D <punnarao.d@lge.com>
Reviewed-by: Rama Krishna <rama.krishna@lge.com>