Add support for handling the new transition states for PMS for Auto
:Release Notes:
PMS needs to support GarageMode state and the transition to GarageMode
Power management service is updated to handle the new states and allow
application to register to and transition to this new state

:Detailed Notes:
A new state is required in between Active state and PowerOff State
during which the application can schedule maintenance and updated
activity. PMS need to implement changes for handling transition to
this new state and the timer logic to decide transition to other
states from GarageMode.

:Testing Performed:
Build Successful

:QA Notes:
NA

:Issues Addressed:
[PLAT-106182] PMS for webOS Auto - Implementation for Garage Mode

Change-Id: I047bb850b3c0c920bb610ba5cc6cbe891b7de76f
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.power2/+/281489
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Punnarao D <punnarao.d@lge.com>
Reviewed-by: Ravi S <ravi.s@lge.com>
Tested-by: Ravi S <ravi.s@lge.com>